Create tower preview class
This creates a class and gets it to load the map in.
