+++ data/maps/tower_test.tmx
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<map version="1.2" tiledversion="1.3.3" orientation="orthogonal" renderorder="right-down" width="12" height="12" tilewidth="42" tileheight="42" infinite="0" nextlayerid="11" nextobjectid="148">
+ <tileset firstgid="1" source="body_tiles.tsx"/>
+ <tileset firstgid="49" source="artery_and_vein.tsx"/>
+ <layer id="1" name="background" width="12" height="12">
+  <data encoding="csv">
+38,42,42,42,42,42,42,42,42,42,42,42,
+30,1,1,1,1,1,1,1,1,1,1,1,
+30,1,1,21,18,18,18,18,18,18,18,18,
+30,1,21,14,13,13,13,13,13,13,13,13,
+30,1,6,13,13,13,13,13,13,13,13,13,
+30,1,6,13,13,13,13,13,13,13,13,13,
+30,1,6,13,13,13,13,13,13,13,13,13,
+30,1,6,13,13,13,13,13,13,13,13,13,
+30,1,6,13,13,13,13,13,13,13,13,13,
+30,1,6,13,13,13,13,13,13,13,13,13,
+30,1,6,13,13,13,13,13,13,13,13,13,
+30,1,6,13,13,13,13,13,13,13,13,13
+</data>
+ </layer>
+ <objectgroup id="4" name="obstacles">
+  <object id="141" name="start" x="42" y="462" width="42" height="42"/>
+  <object id="142" name="goal" x="462" y="42" width="42" height="42"/>
+  <object id="143" name="wall" x="0" y="0" width="42" height="504"/>
+  <object id="144" name="wall" x="42" y="0" width="462" height="42"/>
+  <object id="145" name="wall" x="84" y="126" width="420" height="378"/>
+  <object id="146" name="wall" x="126" y="84" width="378" height="42"/>
+  <object id="147" x="-52.5" y="105"/>
+ </objectgroup>
+</map>
